+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+## Added
+
+- ✨(backend) limit link reach/role select options depending on ancestors #645
+- ✨(backend) add new "descendants" action to document API endpoint #645
+- ✨(backend) new "tree" action on document detail endpoint #645
+- ✨(backend) allow forcing page size within limits #645
+
 ## Changed
 
 - 🛂(frontend) Restore version visibility #629
@@ -17,6 +24,7 @@ and this project adheres to
 ## Fixed
 
 - ♻️(frontend) improve table pdf rendering
+- 🐛(backend) refactor to fix filtering on children and descendants views #645
 
 ## [2.2.0] - 2025-02-10